+++ crates/configuration/src/version4/mod.rs
@@ -3,12 +3,13 @@
 mod comparison;
 pub mod connection_settings;
 pub mod metadata;
+pub mod native_operations;
 mod options;
 mod to_runtime_configuration;
 mod upgrade_from_v3;
 
 use std::borrow::Cow;
-use std::collections::{BTreeMap, HashSet};
+use std::collections::HashSet;
 use std::path::Path;
 pub use to_runtime_configuration::make_runtime_configuration;
 pub use upgrade_from_v3::upgrade_from_v3;
@@ -25,8 +26,6 @@ use metadata::database;
 use crate::environment::Environment;
 use crate::error::{ParseConfigurationError, WriteParsedConfigurationError};
 use crate::values::{ConnectionUri, Secret};
-
-use self::metadata::ScalarTypeName;
 
 #[cfg(test)]
 mod tests;
@@ -76,6 +75,18 @@ impl ParsedConfiguration {
             // we'll change this to `Some(MutationsVersions::V1)` when we
             // want to "release" this behaviour
             mutations_version: None,
+        }
+    }
+
+    /// Extract the connection uri from the configuration + ENV if needed.
+    pub fn get_connection_uri(&self) -> Result<String, anyhow::Error> {
+        let connection_uri = self.connection_settings.connection_uri.clone();
+
+        match connection_uri.0 {
+            super::values::Secret::Plain(connection_string) => Ok(connection_string),
+            super::values::Secret::FromEnvironment { variable } => {
+                Ok(std::env::var(variable.to_string())?)
+            }
         }
     }
 }
@@ -273,73 +284,4 @@ pub async fn write_parsed_configuration(
     .await?;
 
     Ok(())
-}
-
-/// Representation of a result row returned from the oid lookup query.
-#[derive(Debug, sqlx::FromRow)]
-struct OidQueryRow {
-    schema_name: String,
-    type_name: String,
-    oid: i32,
-}
-
-/// Given a vector of OIDs, ask postgres to provide the equivalent type names.
-pub async fn oids_to_typenames(
-    configuration: &ParsedConfiguration,
-    connection_string: &str,
-    oids: &Vec<i64>,
-) -> Result<BTreeMap<i64, ScalarTypeName>, sqlx::Error> {
-    let mut connection = PgConnection::connect(connection_string)
-        .instrument(info_span!("Connect to database"))
-        .await?;
-
-    let rows: Vec<OidQueryRow> = sqlx::query_as(
-        "SELECT
-          typnamespace::regnamespace::text as schema_name,
-          typname as type_name,
-          oid::integer
-        FROM pg_type
-        WHERE oid in (SELECT unnest($1))
-        ",
-    )
-    .bind(oids)
-    .fetch_all(&mut connection)
-    .instrument(info_span!("Run oid lookup query"))
-    .await?;
-
-    let mut oids_map: BTreeMap<i64, ScalarTypeName> = BTreeMap::new();
-
-    // Reverse lookup the schema.typename and find the ndc type name,
-    // if we find all we can just add the nq and call it a day.
-    for row in rows {
-        let schema_name: String = row.schema_name;
-        let type_name: String = row.type_name;
-        let oid: i64 = row.oid.into();
-
-        let mut found = false;
-        for (scalar_type_name, info) in &configuration.metadata.scalar_types.0 {
-            if info.schema_name == schema_name && info.type_name == type_name {
-                oids_map.insert(oid, scalar_type_name.clone());
-                found = true;
-                continue;
-            }
-        }
-
-        // If we don't find it we generate a name which is either schema_typename
-        // or just typename depending if the schema is in the unqualified list or not,
-        // then add the nq and run the introspection.
-        if !found {
-            if configuration
-                .introspection_options
-                .unqualified_schemas_for_types_and_procedures
-                .contains(&schema_name)
-            {
-                oids_map.insert(oid, ScalarTypeName(type_name));
-            } else {
-                oids_map.insert(oid, ScalarTypeName(format!("{schema_name}_{type_name}")));
-            }
-        }
-    }
-
-    Ok(oids_map)
 }
